+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-### [0.5.6]
+### [0.6.0]
 
 New roles:
 - Alpha Werewolf: Can turn other players into werewolves
@@ -7,6 +7,8 @@ New roles:
 Minor additions and changes
 - Added the option to disable the quota from the LethalConfig menu
 
+Bug fixes
+- Fixed a bug that would prevent closing the vote window at the lobby
 
 ### [0.5.5]